requirements.txt
numpy==1.16.5
scikit-image==0.15.0
scipy==1.3.1
scikit-learn==0.21.3
matplotlib==3.1.1
PyQt5==5.9.2
pyyaml==5.1.2